+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CASMCMS-8821 - add support for remote build jobs.
 - CASMCMS-8818 - ssh key injection into jobs.
 - CASMCMS-8897 - changes for aarch64 remote build.
+- CASMCMS-8895 - allow multiple concurrent remote customize jobs.
 
 ## [2.11.0] - 2023-09-15
 ### Changed